Separate AS & QS message processing
