+++ coreclient/src/clients/process/mod.rs
@@ -1,0 +1,12 @@
+// SPDX-FileCopyrightText: 2024 Phoenix R&D GmbH <hello@phnx.im>
+//
+// SPDX-License-Identifier: AGPL-3.0-or-later
+
+use super::{
+    anyhow, AsCredentials, Asset, Contact, ContactAddInfos, Conversation, ConversationAttributes,
+    ConversationId, CoreUser, EarEncryptable, FriendshipPackage, InfraCredentialSigningKey,
+    SignatureEarKey, TimestampedMessage, UserProfile, Verifiable,
+};
+
+pub mod process_as;
+pub mod process_qs;
